Make SequenceWidget 'structural' and give it a FIELDSET.
This make SequenceWidget and MappingWidget behave more consistent.
And: it's already stated in the docstring of Widget: "the default
mapping and sequence widgets are in this category ['structural']".
And: "Renders a sequence into a set of fields."

Plus: we get rid of the invalid LABEL element pointing to the widget's
DIV -- which is not valid HTML.